requirements.txt
Django>=4.2.0
paho-mqtt>=1.6.1 